Refactor TUI application framework into ApplicationHost and DemoApplication
- Renamed TUI Application framework to ApplicationHost
- Extracted reusable application lifecycle functionality into TUI/App/ApplicationHost
- Moved renderer ownership, input processing, event dispatch, resize handling, rendering loop, and screen management into ApplicationHost
- Converted ApplicationHost into a reusable base class for TUI applications
- Created TUI_Demo/App/DemoApplication derived from ApplicationHost
- Moved demo-specific startup behavior out of the framework layer
- Moved demo screen registration and navigation responsibilities into DemoApplication
- Preserved existing demo functionality while reducing framework/demo coupling
- Established framework-owned mechanisms and application-owned policies architecture
- Enabled external applications to derive from ApplicationHost without inheriting demo-specific behavior
- Prepared TUI for use as a reusable framework library by PFF and future applications
