+++ .gitignore
@@ -34,6 +34,22 @@ bld/
 [Ll]og/
 [Ll]ogs/
 
+# Build ouput
+.vs/
+x64/
+Debug/
+Release/
+*.obj
+*.pdb
+*.idb
+*.ilk
+*.lib
+*.exe
+*.log
+*.user
+TUI.vcxproj.user
+TUI_Demo.vcxproj.user
+
 # Visual Studio 2015/2017 cache/options directory
 .vs/
 # Uncomment if you have tasks that create the project's static files in wwwroot
